pbt_gans/requirements.txt
adium-theme-ubuntu==0.3.4
backports.functools-lru-cache==1.4
backports.weakref==1.0.post1
bleach==1.5.0
commentjson==0.6
cycler==0.10.0
decorator==4.0.6
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
h5py==2.7.1
html5lib==0.9999999
Keras==2.2.4
Markdown==2.6.9
matplotlib==2.1.1
mock==2.0.0
nose==1.3.7
numpy==1.14.3
olefile==0.44
opencv-python==3.3.0.10
pandas==0.22.0
pbr==3.1.1
Pillow==4.3.0
prettytensor==0.7.4
protobuf==3.5.0.post1
pydensecrf==1.0rc3
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
scikit-image==0.10.1
scikit-learn==0.19.1
scipy==1.1.0
seaborn==0.8.1
SimpleITK==1.1.0rc1
six==1.11.0
sklearn==0.0
subprocess32==3.2.7
tensorflow-gpu==1.4.0
tensorflow-tensorboard==0.4.0rc3
tensorvision==0.1.dev1
tqdm==4.19.6
unity-lens-photos==1.0
virtualenv==15.1.0
Werkzeug==0.12.2
